refactor(core): Move more code out of NodeExecutionFunctions, and add additional tests (no-changelog) (#13195)
