+++ packages/core/src/execution-engine/node-execution-context/credentials-test-context.ts
@@ -3,8 +3,9 @@ import type { ICredentialTestFunctions } from 'n8n-workflow';
 
 import { Memoized } from '@/decorators';
 import { Logger } from '@/logging';
-// eslint-disable-next-line import/no-cycle
-import { getSSHTunnelFunctions, proxyRequestToAxios } from '@/node-execute-functions';
+
+import { proxyRequestToAxios } from './utils/request-helper-functions';
+import { getSSHTunnelFunctions } from './utils/ssh-tunnel-helper-functions';
 
 export class CredentialTestContext implements ICredentialTestFunctions {
 	readonly helpers: ICredentialTestFunctions['helpers'];
