+++ packages/core/src/execution-engine/node-execution-context/execute-context.ts
@@ -23,16 +23,6 @@ import {
 	NodeConnectionType,
 } from 'n8n-workflow';
 
-// eslint-disable-next-line import/no-cycle
-import {
-	returnJsonArray,
-	copyInputItems,
-	normalizeItems,
-	constructExecutionMetaData,
-	getRequestHelperFunctions,
-	getSSHTunnelFunctions,
-} from '@/node-execute-functions';
-
 import { BaseExecuteContext } from './base-execute-context';
 import {
 	assertBinaryData,
@@ -41,9 +31,15 @@ import {
 	getBinaryHelperFunctions,
 	detectBinaryEncoding,
 } from './utils/binary-helper-functions';
+import { constructExecutionMetaData } from './utils/construct-execution-metadata';
+import { copyInputItems } from './utils/copy-input-items';
 import { getDeduplicationHelperFunctions } from './utils/deduplication-helper-functions';
 import { getFileSystemHelperFunctions } from './utils/file-system-helper-functions';
 import { getInputConnectionData } from './utils/get-input-connection-data';
+import { normalizeItems } from './utils/normalize-items';
+import { getRequestHelperFunctions } from './utils/request-helper-functions';
+import { returnJsonArray } from './utils/return-json-array';
+import { getSSHTunnelFunctions } from './utils/ssh-tunnel-helper-functions';
 
 export class ExecuteContext extends BaseExecuteContext implements IExecuteFunctions {
 	readonly helpers: IExecuteFunctions['helpers'];
