+++ packages/core/src/execution-engine/node-execution-context/hook-context.ts
@@ -11,14 +11,9 @@ import type {
 } from 'n8n-workflow';
 import { ApplicationError } from 'n8n-workflow';
 
-// eslint-disable-next-line import/no-cycle
-import {
-	getNodeWebhookUrl,
-	getRequestHelperFunctions,
-	getWebhookDescription,
-} from '@/node-execute-functions';
-
 import { NodeExecutionContext } from './node-execution-context';
+import { getRequestHelperFunctions } from './utils/request-helper-functions';
+import { getNodeWebhookUrl, getWebhookDescription } from './utils/webhook-helper-functions';
 
 export class HookContext extends NodeExecutionContext implements IHookFunctions {
 	readonly helpers: IHookFunctions['helpers'];
