+++ packages/core/src/execution-engine/node-execution-context/index.ts
@@ -1,18 +1,18 @@
-// eslint-disable-next-line import/no-cycle
 export { CredentialTestContext } from './credentials-test-context';
-// eslint-disable-next-line import/no-cycle
 export { ExecuteContext } from './execute-context';
 export { ExecuteSingleContext } from './execute-single-context';
 export { HookContext } from './hook-context';
 export { LoadOptionsContext } from './load-options-context';
 export { LocalLoadOptionsContext } from './local-load-options-context';
 export { PollContext } from './poll-context';
-// eslint-disable-next-line import/no-cycle
 export { SupplyDataContext } from './supply-data-context';
 export { TriggerContext } from './trigger-context';
 export { WebhookContext } from './webhook-context';
 
+export { constructExecutionMetaData } from './utils/construct-execution-metadata';
 export { getAdditionalKeys } from './utils/get-additional-keys';
+export { normalizeItems } from './utils/normalize-items';
 export { parseIncomingMessage } from './utils/parse-incoming-message';
-export { parseRequestObject } from './utils/parse-request-object';
+export { parseRequestObject } from './utils/request-helper-functions';
+export { returnJsonArray } from './utils/return-json-array';
 export * from './utils/binary-helper-functions';
