+++ packages/core/src/execution-engine/node-execution-context/trigger-context.ts
@@ -9,16 +9,12 @@ import type {
 } from 'n8n-workflow';
 import { ApplicationError, createDeferredPromise } from 'n8n-workflow';
 
-// eslint-disable-next-line import/no-cycle
-import {
-	getRequestHelperFunctions,
-	getSchedulingFunctions,
-	getSSHTunnelFunctions,
-	returnJsonArray,
-} from '@/node-execute-functions';
-
 import { NodeExecutionContext } from './node-execution-context';
 import { getBinaryHelperFunctions } from './utils/binary-helper-functions';
+import { getRequestHelperFunctions } from './utils/request-helper-functions';
+import { returnJsonArray } from './utils/return-json-array';
+import { getSchedulingFunctions } from './utils/scheduling-helper-functions';
+import { getSSHTunnelFunctions } from './utils/ssh-tunnel-helper-functions';
 
 const throwOnEmit = () => {
 	throw new ApplicationError('Overwrite TriggerContext.emit function');
