+++ packages/core/src/execution-engine/node-execution-context/utils/scheduling-helper-functions.ts
@@ -1,0 +1,12 @@
+import { Container } from '@n8n/di';
+import type { SchedulingFunctions, Workflow } from 'n8n-workflow';
+
+import { ScheduledTaskManager } from '../../scheduled-task-manager';
+
+export const getSchedulingFunctions = (workflow: Workflow): SchedulingFunctions => {
+	const scheduledTaskManager = Container.get(ScheduledTaskManager);
+	return {
+		registerCron: (cronExpression, onTick) =>
+			scheduledTaskManager.registerCron(workflow, cronExpression, onTick),
+	};
+};
